+++ HISTORY.md
@@ -3,6 +3,9 @@ https://github.com/stanleyxu2005/dashing
 
 The file only contains the major change history. For full change history, please check the commit log.
 
+## 0.4.7
+* Improvement (#63): Hide placeholder text when search box is focused
+
 ## 0.4.6
 * Feature (#62): Create a confirmation dialog service
 * Improvement (#60): Hide table header when no records in table (not in search mode)